requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
# Flask-Login==0.6.3
# SQLAlchemy==2.0.23
Werkzeug==2.3.7
# python-dotenv==1.0.0
PyJWT==2.8.0